Added MoveSentCheck class and Updated GameScreen
Added class that has method that checks if an array with a size of 2 has both indexes filled. When user sends move using a button, the move is stored into the [0] of array while in the server, the opponents move that they sent is stored in [1] of array. Method checks if array is filled and then calculations on moves is executed.